requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.32.3
shortuuid==1.0.13
urllib3==2.2.2
Werkzeug==3.0.3
win32-setctime==1.1.0
